refactor: LottoResult가 LottoPrize를 반환하게 한다
